cleaning

+++ tp4/src/main/java/fr/ensai/running/controller/AthleteController.java
@@ -19,29 +19,44 @@ public class AthleteController {
     @Autowired
     private AthleteService athleteService;
 
+    /**
+     * Display all Athletes
+     */
     @GetMapping()
     public String findAllAthletes(Model model) {
         model.addAttribute("athletes", athleteService.findAll());
-        return "allAthletes";
+        return "allAthletes"; // allAthlete.html
     }
 
+    /**
+     * Redirection to the add form
+     */
     @GetMapping("/addNew")
     public String addNewAthlete(Model model) {
         Athlete athlete = new Athlete();
         model.addAttribute("athlete", athlete);
-        return "createAthlete";
+        return "createAthlete"; // createAthlete.html
     }
 
+    /**
+     * Save the created Athlete
+     */
     @PostMapping("/save")
     public String saveAthlete(@ModelAttribute("athlete") Athlete athlete) {
         athleteService.save(athlete);
-        return "redirect:/athlete";
+        return "redirect:/athlete"; // redirection to endpoint /athlete
     }
 
+    /**
+     * Delete an Athlete by id
+     */
     @GetMapping("/delete/{id}")
     public String deleteAthleteById(@PathVariable(value = "id") long id) {
-        athleteService.deleteById(id);
-        return "redirect:/athlete";
+        Athlete athlete = athleteService.findById(id);
+        if (athlete != null) {
+            athleteService.deleteById(id);
+        }
+        return "redirect:/athlete"; // redirection to endpoint /athlete
 
     }
 }
